More refactoring

+++ src/Models/Pool.cs
@@ -10,7 +10,7 @@ namespace BakuchiApi.Models
         public Guid EventId { get; set; }
         public string Alias { get; set; }
         public BetType BetType { get; set; }
-        public double TotalWagers { get; set; }
+        public long TotalWagers { get; set; }
         public string Description { get; set; }
 
         // Navigation property
